Everything compiles now.  Added AKILL support.  Also changed lots of files to use C++ style casts as much as possible.
git-svn-id: http://svn.neuromancy.net/magick/trunk/Magick-III@1713 1d68a2a4-f605-0410-a2a0-863aa388aa54

+++ src/server.cpp
@@ -654,7 +654,7 @@ bool Uplink::NUMERIC(Numeric_t num, const std::string &target,
 	MT_EB
 	MT_FUNC("Uplink::NUMERIC" << num << target << args << forcecolon);
 
-	std::string ostr, istr = boost::lexical_cast<std::string>(num) + " " +
+	std::string ostr, istr = boost::lexical_cast<std::string>(static_cast<int>(num)) + " " +
 								target + ROOT->proto.assemble(args, forcecolon);
 	ROOT->proto.addline(*this, ostr, istr);
 	bool rv = ROOT->proto.send(ostr);
